+++ apple-codesign/src/entitlements.rs
@@ -5,7 +5,7 @@
 /*! Code entitlements handling. */
 
 use {
-    crate::AppleCodesignError,
+    crate::{AppleCodesignError, ExecutableSegmentFlags},
     plist::Value,
     rasn::{
         ber::enc::{Encoder as DerEncoder, Error as DerError},
@@ -66,6 +66,52 @@ pub fn der_encode_entitlements_plist(value: &Value) -> Result<Vec<u8>, AppleCode
         })
     })
     .map_err(|e| AppleCodesignError::EntitlementsDerEncode(format!("{}", e)))
+}
+
+/// Convert an entitlements plist to [ExecutableSegmentFlags].
+///
+/// Some entitlements plist values imply features in executable segment flags.
+/// This function resolves those implied features.
+pub fn plist_to_executable_segment_flags(value: &Value) -> ExecutableSegmentFlags {
+    let mut flags = ExecutableSegmentFlags::empty();
+
+    if let Value::Dictionary(d) = value {
+        if matches!(d.get("get-task-allow"), Some(Value::Boolean(true))) {
+            flags |= ExecutableSegmentFlags::ALLOW_UNSIGNED;
+        }
+        if matches!(d.get("run-unsigned-code"), Some(Value::Boolean(true))) {
+            flags |= ExecutableSegmentFlags::ALLOW_UNSIGNED;
+        }
+        if matches!(
+            d.get("com.apple.private.cs.debugger"),
+            Some(Value::Boolean(true))
+        ) {
+            flags |= ExecutableSegmentFlags::DEBUGGER;
+        }
+        if matches!(d.get("dynamic-codesigning"), Some(Value::Boolean(true))) {
+            flags |= ExecutableSegmentFlags::JIT;
+        }
+        if matches!(
+            d.get("com.apple.private.skip-library-validation"),
+            Some(Value::Boolean(true))
+        ) {
+            flags |= ExecutableSegmentFlags::SKIP_LIBRARY_VALIDATION;
+        }
+        if matches!(
+            d.get("com.apple.private.amfi.can-load-cdhash"),
+            Some(Value::Boolean(true))
+        ) {
+            flags |= ExecutableSegmentFlags::CAN_LOAD_CD_HASH;
+        }
+        if matches!(
+            d.get("com.apple.private.amfi.can-execute-cdhash"),
+            Some(Value::Boolean(true))
+        ) {
+            flags |= ExecutableSegmentFlags::CAN_EXEC_CD_HASH;
+        }
+    }
+
+    flags
 }
 
 #[cfg(test)]